+++ packages/docs/docs/main/change-log.md
@@ -11,10 +11,15 @@ This page lists the changes to the IsaacScript framework.
 
 ## July 12th, 2022
 
+- Breaking:
+  - `addRoomClearCharge` now has an argument of `bigRoomDoubleCharge` instead of `ignoreBigRoomDoubleCharge`, so you will need to invert the boolean.
 - All of the `spawn` helper functions now have the option to pass an RNG object instead of a seed.
+- `addRoomClearCharge` and `addRoomClearChargeToSlot` now take an optional argument of `playSoundEffect`.
 - Added the following helper functions:
   - `playerHasHealthLeft`
   - `clearFloorDisplayFlags`
+  - `isRoomShapeDoubleCharge`
+  - `getRoomShapeCharges`
 
 ## July 5th, 2022